Refactor: move deck lists to constants file
